requirements.txt
aiohttp
aiohttp_cors
noise
msgpack
bson
pillow
fabric
boto3
noise
ipython
jinja2
